Extract LevelDB-based token storage from tokendb.go
